+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * added a theme for Bake. See the `README` file;
 * added `IntegrationTestTrait::getStatusCode()` method;
 * fixed a little bug for `HtmlHelper::link()`;
+* `datepicker()`, `datetimepicker()` and `timepicker()` methods provided by
+    `FormHelper` are now deprecated. Use instead the normal `control()` method,
+    which will generate `date`/`datetime-local`/`time` inputs, recognized by the browser;
 * little fixes for `FormHelper`, to make the output more consistent with what
     Bootstrap requires;
 * updated some css file, to make the output more consistent with what Bootstrap